refactor: Delete processor/merger/location_resolver modules
